It is a desktop application for students to test their abilities
It is Desktop Based Application in which user can give test according to their subject. This Application is for two types of users, one is Admin and the other is Student. Admin can add new Subjects and questions as well as can delete subjects and questions.
Student can select a subject and test their skills in that subject. The result will be shown to Student after test as well as emailed to the student at email address specified in personal details of students.